renamed REST module to Client-REST
