+++ Client-REST/src/main/java/pl/jakubowskiprzemyslaw/REST/configuration/RabbitMQConfiguration.java
@@ -4,7 +4,6 @@ import org.springframework.amqp.core.Queue;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import pl.jakubowskiprzemyslaw.REST.models.Player;
 
 @Configuration
 public class RabbitMQConfiguration {
