+++ Client-REST/src/main/java/pl/jakubowskiprzemyslaw/REST/controllers/FleetPlacementController.java
