+++ Client-REST/src/main/java/pl/jakubowskiprzemyslaw/REST/controllers/GameConfigController.java
@@ -8,33 +8,33 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import pl.jakubowskiprzemyslaw.REST.models.Game;
-import pl.jakubowskiprzemyslaw.REST.models.Player;
 import pl.jakubowskiprzemyslaw.REST.services.QueueService;
 import pl.jakubowskiprzemyslaw.REST.services.SessionService;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.validation.Valid;
 
 @Controller
-public class PlayerConfigController extends RESTController {
+public class GameConfigController extends RESTController {
 
-  @Value("${queue.playerConfig}")
-  private String playerConfigQueueName;
+  @Value("${queue.gameConfig}")
+  private String gameConfigQueueName;
 
   @Autowired
-  public PlayerConfigController(QueueService queueService, SessionService sessionService) {
+  public GameConfigController(QueueService queueService, SessionService sessionService) {
     super(queueService, sessionService);
   }
 
-  @GetMapping(value = "/playerconfig", produces = "text/html")
+  @GetMapping(value = "/gameconfig", produces = "text/html")
   public String getGameConfig(Model model) {
-    model.addAttribute("player", new Player());
-    return "playerconfig";
+    model.addAttribute("configuration", new Game());
+    return "gameconfig";
   }
 
-  @PostMapping(value = "/playerconfig", produces = "text/html")
-  public String sendGameConfig(@ModelAttribute("player") Player player, HttpServletRequest request) {
-    addObjectToSession(request, player);
-    sendMessageToQueue(playerConfigQueueName, player);
-    return "redirect:/fleetplacement";
+  @PostMapping(value = "/gameconfig", produces = "text/html")
+  public String sendGameConfig(@Valid @ModelAttribute("configuration") Game configuration, HttpServletRequest request) {
+    addObjectToSession(request, configuration);
+    sendMessageToQueue(gameConfigQueueName, configuration);
+    return "redirect:/playerconfig";
   }
 }
