+++ Client-REST/src/main/java/pl/jakubowskiprzemyslaw/REST/controllers/NoPlacesController.java
