+++ Client-REST/src/main/java/pl/jakubowskiprzemyslaw/REST/services/SessionService.java
