+++ Client-REST/src/main/resources/templates/fleetplacement.html
