+++ Client-REST/src/main/resources/templates/gameconfig.html
