+++ Client-REST/src/main/resources/templates/noplaces.html
