+++ Client-REST/src/main/resources/templates/playerconfig.html
