--- Client/src/main/java/pl/jakubowskiprzemyslaw/tajgertim/queues/QueueSender.java
@@ -1,4 +1,0 @@
-package pl.jakubowskiprzemyslaw.tajgertim.queues;
-
-public class QueueSender {
-}
