--- Client/src/main/java/pl/jakubowskiprzemyslaw/tajgertim/services/QueueService.java
@@ -1,8 +1,0 @@
-package pl.jakubowskiprzemyslaw.tajgertim.services;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class QueueService {
-
-}
