--- Client/src/main/resources/templates/playerconfig.html
@@ -1,17 +1,0 @@
-<!DOCTYPE HTML>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-    <title>BattleSheep game</title>
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
-</head>
-
-<body>
-	<h1>Add Player</h1>
-    <form action="#" th:action="@{/player}" th:object="${player}" method="post">
-    	<p>Name: <input type="text" th:field="*{name}" /></p>
-        <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
-    </form>
-
-
-</body>
-</html>
